% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enar2018pkg.R
\docType{package}
\name{enar2018pkg}
\alias{enar2018pkg}
\alias{enar2018pkg-package}
\title{Infrastructure Tools Package for ENAR 2018}
\description{
This package encapsulates various scripts and templates
we've found to be useful in organizing projects.
}
